requirements.txt
# Core dependencies
pydantic>=2.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# API and networking
aiohttp>=3.9.0
requests>=2.31.0

# Jupyter notebook support
jupyter>=1.0.0
nbformat>=5.9.0
nbconvert>=7.0.0

# Utilities
python-dateutil>=2.8.2
typing-extensions>=4.8.0
pathlib>=1.0.1

# Logging and monitoring
structlog>=23.1.0
prometheus-client>=0.17.0

# Security
cryptography>=41.0.0
python-jose>=3.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0

# Redis for caching
redis>=5.0.1

# React Flow integration
streamlit-component-lib>=1.0.0

# Added pgvector
pgvector

# Added sentence-transformers
sentence-transformers 